Add files via upload
In this project, I designed, debugged and tested a  Java application that manages residents clients and staff information for Ajax Apartments Inc. The application accesses individual or all staff information with great ease, and print out the needed relevant information. This application also manages residents clients of  this same . This is my third project for Ajax Apartments Inc. To see the first and second projects for the same company, view the links on my GitHub.

All rights reserved ®